Migrate chat server to alpine linux
DOCKERFILE
1. image based in alpine

2. we now download our own apk of ejabberd that is hosted in amazon.  we
could get it from the alpine repos, but it can only be downloaded from
the testing branch, which means the apk file was built to support an
'edge' environment.  as we aren't working with edge ejabberd will not
compile properly.  in the meantime, we have to build the apk file to
support alpine stable, download it, and do a local install.

3. install bash as ejabberd requires it (doesn't work with ash)

CONFIG changes
1. Using new config files that come with alpine, ejabberd installation
2. Updated start.sh to work with alpine environment