avoid having 'sh -c "..."' as pid 1 (init) in Docker
Facts:

* Pid 1 is the one that receives SIGTERM when we run `docker stop`.
* Shell scripts swallow signals.
* If the `CMD` sentence in `Dockerfile` is not in the form of a JSON
  array, docker launches a shell to execute the `CMD`, using `sh -c
  "$CMD"`.

So, our pid 1 in the container was a `sh -c "..."` process, that did not
pass signals to `eyeos-run-server`, so `eyeos-run-server` never received
a `SIGTERM` when someone did `docker stop`, and docker ended up killing
our containers in a not-so-graceful way.

To avoid this, we need to specify the `CMD` in "execform" (see
https://docs.docker.com/engine/reference/builder/#cmd) so the pid 1
process in the container is our command instead of a wrapping `sh -c`.
Now, `eyeos-run-server` is the PID 1 and receives the signals when
stopping the container. It then sends SIGTERM to our `start.sh` (because
it is the main child of `eyeos-run-server`) and `start.sh` should finish
ejabberd and then exit so the container can be destroyed.

So, `start.sh`, on receiving the SIGTERM has to stop `ejabberd` and wait
for it to be completely stopped (because it stops in the background).
And when it is finally stopped, we can kill the `tail` process which is
what is making that `start.sh` doesn't end (because of the final
`wait`).